feat: add contracts to umbrella package
