+++ packages/contracts-ts/src/order.ts
@@ -1,84 +1,5 @@
-import { BigIntish, Bytes, getGlobalAdapter, TypedDataDomain, TypedDataTypes } from '@cowprotocol/sdk-common'
-
-/**
- * Gnosis Protocol v2 order data.
- */
-export interface Order {
-  /**
-   * Sell token address.
-   */
-  sellToken: string
-  /**
-   * Buy token address.
-   */
-  buyToken: string
-  /**
-   * An optional address to receive the proceeds of the trade instead of the
-   * owner (i.e. the order signer).
-   */
-  receiver?: string
-  /**
-   * The order sell amount.
-   *
-   * For fill or kill sell orders, this amount represents the exact sell amount
-   * that will be executed in the trade. For fill or kill buy orders, this
-   * amount represents the maximum sell amount that can be executed. For partial
-   * fill orders, this represents a component of the limit price fraction.
-   */
-  sellAmount: BigIntish
-  /**
-   * The order buy amount.
-   *
-   * For fill or kill sell orders, this amount represents the minimum buy amount
-   * that can be executed in the trade. For fill or kill buy orders, this amount
-   * represents the exact buy amount that will be executed. For partial fill
-   * orders, this represents a component of the limit price fraction.
-   */
-  buyAmount: BigIntish
-  /**
-   * The timestamp this order is valid until
-   */
-  validTo: Timestamp
-  /**
-   * Arbitrary application specific data that can be added to an order. This can
-   * also be used to ensure uniqueness between two orders with otherwise the
-   * exact same parameters.
-   */
-  appData: HashLike
-  /**
-   * Fee to give to the protocol.
-   */
-  feeAmount: BigIntish
-  /**
-   * The order kind.
-   */
-  kind: OrderKind
-  /**
-   * Specifies whether or not the order is partially fillable.
-   */
-  partiallyFillable: boolean
-  /**
-   * Specifies how the sell token balance will be withdrawn. It can either be
-   * taken using ERC20 token allowances made directly to the Vault relayer
-   * (default) or using Balancer Vault internal or external balances.
-   */
-  sellTokenBalance?: OrderBalance
-  /**
-   * Specifies how the buy token balance will be paid. It can either be paid
-   * directly in ERC20 tokens (default) in Balancer Vault internal balances.
-   */
-  buyTokenBalance?: OrderBalance
-}
-
-/**
- * Gnosis Protocol v2 order cancellation data.
- */
-export interface OrderCancellations {
-  /**
-   * The unique identifier of the order to be cancelled.
-   */
-  orderUids: Bytes[]
-}
+import { Bytes, getGlobalAdapter, TypedDataDomain, TypedDataTypes } from '@cowprotocol/sdk-common'
+import { Order } from './types'
 
 /**
  * Marker address to indicate that an order is buying Ether.
@@ -103,20 +24,6 @@ export type Timestamp = number | Date
  * A hash-like app data value.
  */
 export type HashLike = Bytes | number
-
-/**
- * Order kind.
- */
-export enum OrderKind {
-  /**
-   * A sell order.
-   */
-  SELL = 'sell',
-  /**
-   * A buy order.
-   */
-  BUY = 'buy',
-}
 
 /**
  * Order balance configuration.
