+++ packages/contracts-ts/src/sign.ts
@@ -1,12 +1,6 @@
-import {
-  Bytes,
-  getGlobalAdapter,
-  SignatureLike,
-  Signer,
-  TypedDataDomain,
-  TypedDataTypes,
-} from '@cowprotocol/sdk-common'
-import { ORDER_TYPE_FIELDS, Order, normalizeOrder } from './order'
+import { Bytes, getGlobalAdapter, Signer, TypedDataDomain, TypedDataTypes } from '@cowprotocol/sdk-common'
+import { ORDER_TYPE_FIELDS, normalizeOrder } from './order'
+import { EcdsaSigningScheme, SigningScheme, Order, EcdsaSignature } from './types'
 
 /**
  * Value returned by a call to `isValidSignature` if the signature was verified
@@ -14,55 +8,6 @@ import { ORDER_TYPE_FIELDS, Order, normalizeOrder } from './order'
  * bytes4(keccak256("isValidSignature(bytes32,bytes)"))
  */
 export const EIP1271_MAGICVALUE = '0x1626ba7e'
-
-/**
- * The signing scheme used to sign the order.
- */
-export enum SigningScheme {
-  /**
-   * The EIP-712 typed data signing scheme. This is the preferred scheme as it
-   * provides more infomation to wallets performing the signature on the data
-   * being signed.
-   *
-   * <https://github.com/ethereum/EIPs/blob/master/EIPS/eip-712.md#definition-of-domainseparator>
-   */
-  EIP712 = 0b00,
-  /**
-   * Message signed using eth_sign RPC call.
-   */
-  ETHSIGN = 0b01,
-  /**
-   * Smart contract signatures as defined in EIP-1271.
-   *
-   * <https://eips.ethereum.org/EIPS/eip-1271>
-   */
-  EIP1271 = 0b10,
-  /**
-   * Pre-signed order.
-   */
-  PRESIGN = 0b11,
-}
-
-export type EcdsaSigningScheme = SigningScheme.EIP712 | SigningScheme.ETHSIGN
-
-/**
- * The signature of an order.
- */
-export type Signature = EcdsaSignature | Eip1271Signature | PreSignSignature
-
-/**
- * ECDSA signature of an order.
- */
-export interface EcdsaSignature {
-  /**
-   * The signing scheme used in the signature.
-   */
-  scheme: EcdsaSigningScheme
-  /**
-   * The ECDSA signature.
-   */
-  data: SignatureLike
-}
 
 /**
  * EIP-1271 signature data.
