+++ packages/contracts-ts/src/swap.ts
@@ -6,9 +6,9 @@ import {
   Signer,
   TypedDataDomain,
 } from '@cowprotocol/sdk-common'
-import { Order, OrderKind } from './order'
-import { TokenRegistry, Trade, encodeTrade } from './settlement'
-import { EcdsaSigningScheme, Signature, signOrder } from './sign'
+import { TokenRegistry, encodeTrade } from './settlement'
+import { signOrder } from './sign'
+import { Signature, EcdsaSigningScheme, Trade, Order, OrderKind } from './types'
 
 /**
  * A Balancer swap used for settling a single order against Balancer pools.
